requirements.txt
Flask==2.0.1
kafka-python==2.0.2
pandas==1.2.4
numpy==1.20.3
scikit-learn==0.24.2
pycryptodome==3.11.0
psycopg2==2.9.1